+++ src/Exception/NotAnArrayException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace PascaleBeier\Config\Exception;
-
-class NotAnArrayException extends \Exception
-{
-    protected $message = 'The included Configuration File did not return an array.';
-}