+++ tests/config/app.php
@@ -1,0 +1,8 @@
+<?php
+
+return [
+    'name' => 'foo',
+    'software' => [
+        'version' => 'v2.0',
+    ],
+];